Remove SpeechToTextConstants
